rename filename to forward_scanner
